heliot_cdss/pyproject.toml
[tool.poetry]
name = "cdss"
version = "0.1.0"
description = ""
authors = ["Gabriele De Vito <gabriele.devito@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tiledb-vector-search = "^0.3.0"
lxml = "^5.2.1"
bs4 = "^0.0.2"
tiktoken = "^0.6.0"
openai = "^1.28.1"
openpyxl = "^3.1.2"
tiledb = "^0.29.0"
unidecode = "^1.3.8"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
streamlit = "^1.36.0"
streamlit-searchbox = "^0.1.20"
seaborn = "^0.13.2"
nltk = "^3.9.1"
torch = "^2.7.1"
torchvision = "^0.22.1"
torchaudio = "^2.7.1"
transformers = "^4.52.4"
accelerate = "^1.7.0"
tokenizers = "^0.21.1"
anthropic = "^0.53.0"
sqlalchemy = "^2.0.47"
alembic = "^1.18.4"
psycopg = {extras = ["binary"], version = "^3.3.3"}
python-dotenv = "^1.2.2"
redis = "^7.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
